Rename cache -> file + split off validate + mypy
